Add `Content-Length: 0` request header for empty POST etc.
Add documentation for when this library will automatically add an
outgoing `Content-Length` request header and that this header should
only be passed manually for request streaming.

Additionally, we now make sure to include a `Content-Length: 0` request
header if the request body is known to be empty and the request method
usually expects a request body (only applies to `POST`, `PUT` and
`PATCH`). While not strictly required by the newer RFC 7230, this does
improve compatibility with some HTTP servers which seem to follow the
older RFC 2616 semantics.